Reorganize plot module
