+++ .pre-commit-config.yaml
@@ -72,7 +72,7 @@ repos:
       - id: black
 
   - repo: https://github.com/keewis/blackdoc
-    rev: v0.3.9
+    rev: v0.4.0
     hooks:
       - id: blackdoc
         additional_dependencies: ["black==25.1.0"]
@@ -85,13 +85,13 @@ repos:
 #     additional_dependencies: [flake8-typing-imports, flake8-docstrings]
 
 # - repo: https://github.com/astral-sh/ruff-pre-commit
-#   rev: v0.11.12
+#   rev: v0.12.0
 #   hooks:
 #   - id: ruff
 #     args: [--show-fixes]
 
 # - repo: https://github.com/pre-commit/mirrors-mypy
-#   rev: v1.16.0
+#   rev: v1.16.1
 #   hooks:
 #   - id: mypy
 #     files: src
